slash_command_generator/pyproject.toml
[project]
name = "v3-command-generator"
version = "0.1.0"
description = "v3プロジェクト企画フェーズのスラッシュコマンド生成ツール"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]